avoid deleting a category with products inside

+++ src/categories/categories.service.ts
@@ -1,5 +1,5 @@
 /* eslint-disable prettier/prettier */
-import { Injectable, forwardRef, Inject, HttpException, HttpStatus, NotFoundException } from "@nestjs/common";
+import { Injectable, forwardRef, Inject, HttpException, HttpStatus, NotFoundException, ConflictException } from "@nestjs/common";
 import { ProductsService } from "../products/products.service";
 import { Repository } from 'typeorm';
 import { Category } from './entities/category.entity';
@@ -83,15 +83,20 @@ export class CategoriesService {
   async delete(id: number) {
     const category = await this.categoryRepo.findOneBy({ id: id });
     if (!category) {
-      throw new NotFoundException('Categoria no encontrada :(');
+      throw new NotFoundException('Categoría no encontrada :(');
     }
-    else {
-      await this.adjustSequence();
-      this.categoryRepo.delete(id);
-      await this.adjustSequence();
-      return { message: "Categoria borrada" }
-    }
+    const productsInCategory = await this.productsService.filterByCategoryId(id);
+    if (productsInCategory.length > 0) {
+      throw new ConflictException('No se puede borrar la categoría, contiene productos.');
+    } 
+    await this.adjustSequence();
+    await this.categoryRepo.delete(id);
+    await this.adjustSequence();
+  
+    return { message: 'Categoría borrada' };
   }
+  
+  
 
   async update(id: number, updateCategoryDto: UpdateCategoryDto) {
     const category = await this.categoryRepo.findOneBy({ id: id });
